# frontend/semmelei/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Unnamed Galvyn API",
        "version": "v0.0.0"
    },
    "paths": {
        "/api/frontend/v1/auth/login/start": {
            "post": {
                "summary": "Start a passkey login for a given username",
                "description": "Start a passkey login for a given username\n\nThe account's passkeys are sent as the credential allow-list, so this\nworks with roaming authenticators (e.g. YubiKeys) whose credentials are\nnot client-side discoverable.",
                "operationId": "start_login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/login/finish": {
            "post": {
                "summary": "Finish a passkey login",
                "description": "Finish a passkey login\n\nVerifies the browser's credential and logs the account in.",
                "operationId": "finish_login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FinishLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/register/start": {
            "post": {
                "summary": "Start an invite-based passkey registration",
                "description": "Start an invite-based passkey registration",
                "operationId": "start_registration",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartRegistrationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/register/finish": {
            "post": {
                "summary": "Finish an invite-based passkey registration",
                "description": "Finish an invite-based passkey registration",
                "operationId": "finish_registration",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FinishRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/logout": {
            "get": {
                "summary": "Log out of the current session",
                "description": "Log out of the current session",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/frontend/v1/auth/me": {
            "get": {
                "summary": "Get the currently logged-in account",
                "description": "Get the currently logged-in account",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/passkeys": {
            "get": {
                "summary": "List the passkeys of the logged-in account",
                "description": "List the passkeys of the logged-in account",
                "operationId": "list_passkeys",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListPasskeysResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/passkeys/start": {
            "post": {
                "summary": "Start adding another passkey to the logged-in account",
                "description": "Start adding another passkey to the logged-in account",
                "operationId": "start_add_passkey",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartAddPasskeyResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/passkeys/finish": {
            "post": {
                "summary": "Finish adding another passkey to the logged-in account",
                "description": "Finish adding another passkey to the logged-in account",
                "operationId": "finish_add_passkey",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FinishAddPasskeyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "description": "The response that is sent in a case of an error the caller should report to an admin",
                "type": "object",
                "properties": {
                    "trace_id": {
                        "description": "ID of the opentelemetry trace this error originated in",
                        "type": "string"
                    }
                },
                "required": [
                    "trace_id"
                ]
            },
...
            "FinishLoginRequest": {
                "description": "Request to finish a login ceremony",
                "type": "object",
                "properties": {
                    "credential": {
                        "description": "The browser's `PublicKeyCredential` response"
                    }
                },
                "required": [
                    "credential"
                ]
            },
...
            "ListPasskeysResponse": {
                "description": "All passkeys of the current account",
                "type": "object",
                "properties": {
                    "passkeys": {
                        "description": "The passkeys",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PasskeySchema"
                        }
                    }
                },
                "required": [
                    "passkeys"
                ]
            },
...
            "StartLoginRequest": {
                "description": "Request to start a login ceremony",
                "type": "object",
                "properties": {
                    "username": {
                        "description": "The username to authenticate",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/MaxStr_255"
                            }
                        ]
                    }
                },
                "required": [
                    "username"
                ]
            },
            "StartLoginResponse": {
                "description": "Response to a started login ceremony",
                "type": "object",
                "properties": {
                    "options": {
                        "description": "`PublicKeyCredentialRequestOptions` to pass to the browser"
                    }
                },
                "required": [
                    "options"
                ]
            },
...
        }
    }
}